// packages/geometry/src/generated/profileAssets.ts
// Generated by scripts/generate-geometry-profile-assets.ts. Do not edit by hand.
export type GeneratedProfileAsset = {
  widthMm: number;
  depthMm: number;
  sectionOutline: Array<{ x: number; y: number }>;
  sectionVoids: Array<Array<{ x: number; y: number }>> | null;
};

export const GENERATED_PROFILE_ASSETS = {
  sp_gutter: {
    widthMm: 100,
    depthMm: 150.568624,
    sectionOutline: [{ x: -24.003203, y: -49.284312 }, { x: -24.003203, y: 73.009886 }, { x: -50, y: 75.284312 }, { x: -50, y: -75.284312 }, { x: 50, y: -75.284312 }, { x: 50, y: 74.715688 }, { x: 35, y: 74.715688 }, { x: 35, y: 66.715688 }, { x: 38, y: 66.715688 }, { x: 38, y: 71.715688 }, { x: 47, y: 71.715688 }, { x: 47, y: -49.284312 }],
    sectionVoids: [[{ x: -24, y: -72.284312 }, { x: 47, y: -72.284312 }, { x: 47, y: -52.284312 }, { x: -24, y: -52.284312 }], [{ x: -47, y: -72.284312 }, { x: -27, y: -72.284312 }, { x: -27, y: 70.260613 }, { x: -47, y: 72.010386 }]],
  },
  sp_joiners: {
    widthMm: 50,
    depthMm: 16,
    sectionOutline: [{ x: 25, y: -4.5 }, { x: 25, y: -8 }, { x: -25, y: -8 }, { x: -25, y: -4.5 }, { x: -15.95, y: -4.5 }, { x: -15.95, y: -6 }, { x: -7.36, y: -6 }, { x: -7.36, y: 6 }, { x: -15.95, y: 6 }, { x: -15.95, y: 4.5 }, { x: -25, y: 4.5 }, { x: -25, y: 8 }, { x: 25, y: 8 }, { x: 25, y: 4.5 }, { x: 15.95, y: 4.5 }, { x: 15.95, y: 6 }, { x: 7.36, y: 6 }, { x: 7.36, y: -6 }, { x: 15.95, y: -6 }, { x: 15.95, y: -4.5 }],
    sectionVoids: null,
  },
} as const satisfies Record<string, GeneratedProfileAsset>;